requirements.txt
colorlog
python-dotenv
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
langchain
langchain-core
langchain-openai
openai
pydantic
quart
quart-cors
beautifulsoup4
Pillow
requests
termcolor
pytz
PyJWT